requirements.txt
torch
torchvision
numpy
pillow
tqdm
tensorboard
pyyaml
ffmpeg-python